Remove dead code, tweak apis.
