+++ lib/json_api/formatters/json_api_org.ex
@@ -1,28 +1,25 @@
+#TODO move file & directory to reflect singular module name.
 defmodule JsonApi.Formatter.JsonApiOrg do
   alias JsonApi.Formatter.JsonApiOrg.Format
-  alias JsonApi.Formatter.JsonApiOrg.Parser
-  alias JsonApi.Serializer.Attributes
-  alias JsonApi.Serializer.Relationships
+  alias JsonApi.Formatter.JsonApiOrg.Parse
 
   @moduledoc """
   Formats and parses between the jsonapi.org format and Elixir maps.
   """
 
-  #TODO: Should impliment Formatter callback.
-  @doc "Convience function for JsonApiOrg.Format.format"
+  @doc "Convenience function for JsonApiOrg.Format.format"
   def format(model, serializer, conn, meta) do
     Format.format(model, serializer, conn, meta)
   end
 
-  @doc "Convience function for JsonApiOrg.Parse.parse"
-  def parse() do
-    Parser.parse()
+  @doc "Convenience function for JsonApiOrg.Parse.parse"
+  def parse(raw) do
+    Parse.parse(raw)
   end
 
-
   defmodule Parse do
-    def parse do
-
+    def parse(raw) do
+      JsonApi.ConvertKeys.underscore(raw)
     end
   end
 end
